requirements.txt
uvicorn
fastapi
pydantic
pydantic-settings
httpx
loguru
orjson
motor
motor[srv]
markdown2[all]
pillow
jsonpath-ng
Jinja2
python-dotenv